Basic support for submission creation, listing
Very timidly wading into the functional waters of Scala.

Most of this is quickly gleaned from Play and Anorm docs, a
helpful Stack Overflow answer[1], and various sample Play apps on
GitHub.

Seeing it *actually work* is pretty cool! The syntax is a little
over my head, but I'll pick up on it quickly I think.

[1]: http://stackoverflow.com/a/17041347